requirements.txt
absl-py==2.0.0
ale-py==0.8.1
arrow==1.3.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
baselines==0.1.5
box2d-py==2.3.5
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
codecarbon==2.3.2
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
contourpy==1.2.0
cycler==0.12.1
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deepdiff==6.5.0
dill==0.3.7
dm-env==1.6
dm-tree==0.1.8
enum34==1.1.10
exceptiongroup @ file:///C:/b/abs_25wqfvkf25/croot/exceptiongroup_1668714345637/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
Farama-Notifications==0.0.4
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.44.0
fsspec==2023.10.0
future==0.18.3
fuzzywuzzy==0.18.0
gast==0.5.4
glfw==2.6.2
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.9.0
idna==3.4
imageio==2.31.4
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
importlib-resources==6.1.0
ipykernel @ file:///D:/bld/ipykernel_1698244157926/work
ipython @ file:///C:/b/abs_e5729i179y/croot/ipython_1694181400005/work
jedi @ file:///C:/ci/jedi_1644315428289/work
Jinja2==3.1.2
joblib==1.3.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1698231991356/work
jupyter_core @ file:///D:/bld/jupyter_core_1698673840402/work
keras==2.14.0
kiwisolver==1.4.5
lamp==0.0.1
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661915841596/work
ml-dtypes==0.2.0
mock==5.1.0
mpi4py==3.1.4
mpmath==1.3.0
mpyq==0.2.5
mujoco==2.2.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1697083700168/work
networkx==3.2.1
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
ordered-set==4.1.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.1.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==10.0.1
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
portpicker==1.6.0
progressbar2==4.2.0
prometheus-client==0.19.0
prompt-toolkit @ file:///C:/b/abs_6coz5_9f2s/croot/prompt-toolkit_1672387908312/work
protobuf==3.20.0
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
pyaml==23.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.5.2
Pygments @ file:///C:/b/abs_fay9dpq4n_/croot/pygments_1684279990574/work
pynvml==11.5.0
PyOpenGL==3.1.7
pyparsing==3.1.1
PySC2==4.0.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-utils==3.8.1
pytz==2023.3.post1
PyVirtualDisplay==3.0
pywin32==305.1
PyYAML==6.0.1
pyzmq @ file:///D:/bld/pyzmq_1660329100686/work
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s2clientprotocol==5.0.11.90870.0
s2protocol==5.0.11.90870.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sk-video==1.1.10
stable-baselines==2.10.2
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
swig==4.1.1.post0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.22.1
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
torchvision==0.16.1+cu121
tornado @ file:///D:/bld/tornado_1656937934674/work
tqdm==4.66.1
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
types-python-dateutil==2.8.19.14
typing_extensions @ file:///C:/b/abs_213vg2cd1l/croot/typing_extensions_1690297804941/work
tzdata==2023.3
urllib3==2.0.5
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websocket-client==1.6.3
Werkzeug==3.0.0
wrapt==1.14.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
zmq==0.0.0
